pyproject.toml
[project]
name = "cyberteller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyswisseph>=2.10.3.2",
    "langchain-core>=0.1.0",
    "langchain>=0.1.0",
    "langchain-community==0.3.14",
    "langchain-openai==0.2.14",
    "langchain-qdrant==0.2.0",
    "qdrant-client==1.13.2",
    "tiktoken>=0.8.0",
    "pymupdf==1.25.2",
    "langgraph==0.2.61",
    "langsmith>=0.0.52",
    "openai>=1.12.0",
    "cohere==5.13.11",
    "lxml==5.3.0",
    "jupyter==1.1.1",
    "bs4==0.0.2",
    "pydantic>=2.7.4,<3.0.0",
    "grandalf>=0.8",
    "pandas>=2.2.3",
    "pyppeteer>=2.0.0",
    "pypdf2>=3.0.1",
    "geopy>=2.4.1",
    "matplotlib>=3.10.3",
    "timezonefinder>=6.5.9",
    "pytz>=2023.3",
    "nltk>=3.9.1",
    "unstructured[pdf]>=0.17.2",
    "unstructured>=0.14.8",
    "ragas>=0.2.10",
    "anyio>=3.7.1",
    "langchain-huggingface>=0.2.0",
]
